+++ cmd/root.go
@@ -18,7 +18,7 @@ import (
 var (
 	cfgFile string
 	envName string
-	Version = "2.5.1"
+	Version = "2.5.2"
 )
 
 func showBanner() {